Create feature stories
