+++ packages/react/src/Link/Link.stories.tsx
@@ -1,5 +1,5 @@
-import {ComponentMeta, ComponentStory} from '@storybook/react'
 import React from 'react'
+import {ComponentMeta, ComponentStory} from '@storybook/react'
 import {Link} from '.'
 
 export default {
@@ -11,23 +11,7 @@ const Template: ComponentStory<typeof Link> = args => <Link {...args} />
 
 export const Default = Template.bind({})
 Default.args = {
-  size: 'medium',
+  // size: 'medium',
   href: '#',
   children: 'Primary action',
-}
-
-export const Reversed = Template.bind({})
-Reversed.args = {
-  size: 'medium',
-  href: '#',
-  children: 'Back to schedule',
-  arrowDirection: 'start',
-}
-
-export const Accent = Template.bind({})
-Accent.args = {
-  size: 'medium',
-  variant: 'accent',
-  href: '#',
-  children: 'Accent color',
 }
